Dedicated object modules by object type
